infrastructure/cdk/cdk.out/asset.0d64842dfb6c5231587d80bcb91dc10cd2b81b28db0198aada81e593f72497e4/infrastructure/cdk/requirements.txt
aws-cdk-lib==2.236.0
constructs==10.4.4
pyyaml
